requirements.txt
matplotlib==3.10.3
numpy==2.2.5
Pillow==11.2.1
pygame==2.6.1
torch==2.6.0
torchinfo==1.8.0
tqdm==4.67.1
